Revert ID columns to Int32

+++ Hangfire.FluentNHibernateStorage/Entities/_AggregatedCounter.cs
@@ -1,6 +1,6 @@
 ﻿namespace Hangfire.FluentNHibernateStorage.Entities
 {
-    public class _AggregatedCounter : KeyValueTypeBase<long>
+    public class _AggregatedCounter : KeyValueTypeBase<int>
     {
     }
 }
