+++ Hangfire.FluentNHibernateStorage/Entities/_Counter.cs
@@ -1,6 +1,6 @@
 ﻿namespace Hangfire.FluentNHibernateStorage.Entities
 {
-    public class _Counter : KeyValueTypeBase<long>
+    public class _Counter : KeyValueTypeBase<int>
     {
     }
 }
